+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Chip/Filter/FilterChipElement.swift
@@ -25,7 +25,7 @@ struct FilterChipElement: DesignToolboxElement {
         pageDescription = AnyView(DesignToolboxElementPage(
             name: name,
             description: "app_components_chip_filterChip_description_text",
-            version: OUDSVersions.componentChipVersion,
+            version: OUDSVersions.componentFilterChipVersion,
             demoScreen: AnyView(FilterChipPage())))
     }
 }